ml_trainer_container/requirements.txt
tensorflow-gpu
sklearn
mysql-connector-python
google-cloud-storage
